requirements.txt
python-dotenv
fastapi
langgraph 
langchain
langchain_core
streamlit
langchain_groq
fastapi
uvicorn
gunicorn
sqlalchemy 
pymysql